% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_tables.R
\name{addSourceToTarget}
\alias{addSourceToTarget}
\title{title here}
\usage{
addSourceToTarget(sourceSuffix, targetSuffix, dryRun, con)
}
\arguments{
\item{sourceSuffix}{String for suffix of names of source tables. Cannot be
\code{NULL} or ''.}

\item{targetSuffix}{String for suffix of names of target tables.}

\item{dryRun}{Logical indicating whether to perform a dry run to determine
how many rows in each table would be deleted and inserted, without making
any changes to the database.}

\item{con}{Connection to the database, created using \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
A data.table with columns \code{source_name}, \code{target_name},
\code{nrows_delete}, and \code{nrows_insert}. If \code{dryRun} is \code{FALSE}, a side effect
in the database is that some rows in the target tables will be deleted,
some rows in the source tables will be appended to the target tables, and
the source tables will be deleted.
}
\description{
description here
}
