% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPgParams}
\alias{getPgParams}
\title{Get Postgres connection parameters}
\usage{
getPgParams(path = "~/.pgpass")
}
\arguments{
\item{path}{Path to .pgpass file. The file's first line should be
\preformatted{# hostname:port:database:username:password}}
}
\value{
A data.table with one row for each set of parameters.
}
\description{
This is a helper function to get parameters from a .pgpass file. See
\href{https://www.postgresql.org/docs/9.6/libpq-pgpass.html}{here} for details.
}
