% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{local_folder}
\alias{local_folder}
\title{Copy of withr::local_file function, but allows additional parameters.}
\usage{
local_folder(file, ..., .local_envir = parent.frame())
}
\arguments{
\item{file}{\verb{[named list]}\cr Files to create.}

\item{.local_envir}{\verb{[environment]}\cr The environment to use for scoping.}
}
\description{
Allows more parameters to be provided to the unlink step in the withr::local_file function.
}
