% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_tables.R
\name{modifyTables}
\alias{modifyTables}
\title{Create or update a PubMed database}
\usage{
modifyTables(
  localDir,
  dbname,
  dbtype = "postgres",
  nFiles = Inf,
  retry = TRUE,
  nCitations = Inf,
  mode = c("create", "update"),
  ...
)
}
\arguments{
\item{localDir}{Path to directory in which to download the files from PubMed.}

\item{dbname}{Name of database.}

\item{dbtype}{Type of database, either 'postgres', 'mariadb', 'mysql', or
'sqlite'.}

\item{nFiles}{Maximum number of xml files to parse that are not already in
the database. This should not normally be changed from the default.}

\item{retry}{Logical indicating whether to retry parsing steps that fail.}

\item{nCitations}{Maximum number of rows of the citation file to read. This
should not normally be changed from the default.}

\item{mode}{String indicating whether to create the database using the
baseline files or to update the database using the update files.}

\item{...}{Other arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
\code{NULL}, invisibly. Log files will be created in \code{localDir}.
}
\description{
description here
}
