% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pubmed_xml.R
\name{parsePubmedXml}
\alias{parsePubmedXml}
\title{Parse PubMed XML files}
\usage{
parsePubmedXml(
  xmlDir,
  xmlFiles = NULL,
  logPath = NULL,
  tableSuffix = NULL,
  overwrite = FALSE,
  dbtype = "postgres",
  dbname = NULL,
  ...
)
}
\arguments{
\item{xmlDir}{Path to directory containing the xml or xml.gz files.}

\item{xmlFiles}{Character vector of file names or a data.frame with columns
\code{xml_filename} and \code{step} TODO}

\item{logPath}{Path to the log file to create. The log file is a
tab-delimited file with columns \code{datetime}, \code{xml_filename}, \code{step},
\code{status}, and \code{message}. A \code{status} of 0 indicates success, 1 indicates
an error, in which case \code{message} contains the error message.}

\item{tableSuffix}{String to append to the table names.}

\item{overwrite}{Logical indicating whether to overwrite existing tables.}

\item{dbtype}{String indicating type of database, either 'postgres',
'mariadb', 'mysql', or 'sqlite'. Only used if \code{dbname} is not \code{NULL}.}

\item{dbname}{Name of the database in which to create the tables.}

\item{...}{Other arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
describe shit here
}
