% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_element.R, R/parse_person_affiliation.R
\name{parseElement}
\alias{parseElement}
\alias{parsePmidStatus}
\alias{parseArticleId}
\alias{parsePubDate}
\alias{parseTitleJournal}
\alias{parsePubType}
\alias{parseMeshTerm}
\alias{parseKeyword}
\alias{parseGrant}
\alias{parseChemical}
\alias{parseDataBank}
\alias{parseComment}
\alias{parseAbstract}
\alias{parseAuthorAffiliation}
\alias{parseInvestigatorAffiliation}
\title{Parse elements from a PubMed XML file}
\usage{
parsePmidStatus(rawXml, filename, con = NULL, tableSuffix = NULL)

parseArticleId(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parsePubDate(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseTitleJournal(
  pmXml,
  pmids,
  filename = NULL,
  con = NULL,
  tableSuffix = NULL
)

parsePubType(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseMeshTerm(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseKeyword(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseGrant(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseChemical(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseDataBank(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseComment(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseAbstract(pmXml, pmids, filename = NULL, con = NULL, tableSuffix = NULL)

parseAuthorAffiliation(
  pmXml,
  pmids,
  filename = NULL,
  con = NULL,
  tableSuffix = NULL
)

parseInvestigatorAffiliation(
  pmXml,
  pmids,
  filename = NULL,
  con = NULL,
  tableSuffix = NULL
)
}
\arguments{
\item{rawXml}{An xml document obtained by loading a PubMed XML file using
\code{\link[xml2:read_xml]{xml2::read_xml()}}.}

\item{filename}{A string or NULL. If not NULL, a column \code{xml_filename} is
added to the generated table(s).}

\item{con}{Connection to the database, created using \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{tableSuffix}{String to append to the table names.}

\item{pmXml}{An xml nodeset derived from \code{rawXml}, such as that returned by
\code{parsePmidStatus()}, where each node corresponds to a PMID.}

\item{pmids}{Integer vector of PMIDs for \code{pmXml}.}
}
\value{
\code{parsePmidStatus()} returns a list of two objects: the first is an xml
nodeset in which each node corresponds to a PMID, the second is a
data.table with columns \code{pmid} and \code{status}. The latter is parsed from the
DeleteCitation and MedlineCitation sections.

\code{parseArticleId()} returns a data.table with columns \code{pmid}, \code{id_type}, and
\code{id_value}, parsed from the ArticleIdList section. Only \code{id_type}s "doi"
and "pmc" are retained.

\code{parsePubDate()} returns a data.table with columns \code{pmid}, \code{pub_status},
and \code{pub_date}, parsed from the History section.

\code{parseTitleJournal()} returns a data.table with columns \code{pmid}, \code{title},
\code{journal_full}, and \code{journal_abbrev}, parsed from the Journal section.

\code{parsePubType()} returns a data.table with columns \code{pmid}, \code{type_name}, and
\code{type_id}, parsed from the PublicationTypeList section.

\code{parseMeshTerm()} returns a data.table with columns \code{pmid}, \code{term_name},
\code{term_id}, and \code{major_topic}, parsed from the MeshHeadingList section.

\code{parseKeyword()} returns a list of two data.tables: the first has columns
\code{pmid} and \code{list_owner}, the second has columns \code{pmid}, \code{keyword_name},
and \code{major_topic}. Both data.tables are parsed from the KeywordList
section.

\code{parseGrant()} returns a list of two data.tables: the first has columns
\code{pmid} and \code{complete}, the second has columns \code{pmid}, \code{grant_id},
\code{acronym}, \code{agency}, and \code{country}. Both data.tables are parsed from the
GrantList section.

\code{parseChemical()} returns a data.table with columns \code{pmid},
\code{registry_number}, \code{substance_name}, and \code{substance_ui}, parsed from the
ChemicalList section.

\code{parseDataBank()} returns a data.table with columns \code{pmid},
\code{data_bank_name}, and \code{accession_number}, parsed from the DataBankList
section.

\code{parseComment()} returns a data.table with columns \code{pmid}, \code{ref_type}, and
\code{ref_pmid}, parsed from the CommentsCorrectionsList section.

\code{parseAbstract()} returns a data.table with columns \code{pmid}, \code{text}, \code{label},
\code{nlm_category}, and \code{copyright}, parsed from the Abstract section.

\code{parseAuthorAffiliation()} returns a list of data.tables parsed from the
AuthorList section. The first data.table is for authors and has columns
\code{pmid}, \code{author_pos}, \code{last_name}, \code{fore_name}, \code{initials}, \code{suffix},
and \code{collective_name}. The second data.table is for affiliations and has
columns \code{pmid}, \code{author_pos}, \code{affiliation_pos}, and \code{affiliation}. The
third data.table is for author identifiers and has columns \code{pmid},
\code{author_pos}, \code{source}, and \code{identifier}. The fourth data.table is for
author affiliation identifiers and has columns \code{pmid}, \code{author_pos},
\code{affiliation_pos}, \code{source}, and \code{identifier}.

\code{parseInvestigatorAffiliation()} returns a list of data.tables identical to
those returned by \code{parseAuthorAffiliation()}, except parsed from the
InvestigatorList section, with column names containing "investigator"
instead of "author", and lacking a column in the first data.table for
\code{collective_name}.
}
\description{
Elements are parsed according to the MEDLINE®PubMed® XML Element
Descriptions and their Attributes
\href{https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html}{here}.
}
